+++ back/Endpoints.md
@@ -72,11 +72,7 @@ headers:{
 - Mostrar un estudiante
 - El params ':id' es el id del estudiante
 
-### URL: 'http://localhost:8000/admin/advisers'
-- Metodo http: 'GET'
-- Mostrar a todos los orientadores
-
-### URL: 'http://localhost:8000/admin/advisers/student/:id'
+### URL: 'http://localhost:8000/admin/students/:id/adviser'
 - Metodo http: 'PUT'
 - Asignar un orientador a un estudiante
 - El params ':id' es el id del estudiante
@@ -86,6 +82,11 @@ headers:{
     "idAdviser":"3"
 }
 ```
+
+### URL: 'http://localhost:8000/admin/advisers'
+- Metodo http: 'GET'
+- Mostrar a todos los orientadores
+
 ### URL: 'http://localhost:8000/admin/events'
 - Metodo http: 'POST'.
 - Crear un evento
