+++ back/controllers/admin/adviser.js
@@ -1,4 +1,4 @@
-const { Advisers: AdviserModel, Students: StudentModel } = require('../../data/models');
+const { Advisers: AdviserModel } = require('../../data/models');
 const logger = require('../../utils/logger');
 // Method to get all the advisers
 const getAllAdvisers = async (req, res) => {
@@ -15,39 +15,7 @@ const getAllAdvisers = async (req, res) => {
     logger.error(error);
   }
 };
-const assignAdviser = async (req, res) => {
-  try {
-    const { idAdviser } = req.body;
-    const { id } = req.params;
-    const student = await StudentModel.update(
-      {
-        adviserId: idAdviser,
-      },
-      {
-        where: {
-          id,
-        },
-      },
-    );
-    if (student < 1) {
-      logger.warn(`ID Student: '${id}', not found in db. Method: PUT.  Url: ${req.originalUrl}.`);
-      return res.status(204).json({
-        message: 'Student not found',
-        data: '',
-      });
-    }
-    res.status(201).json({
-      message: `Adviser assigned to the student with the id ${id}`,
-    });
-  } catch (error) {
-    res.status(500).json({
-      message: 'Server Error',
-    });
-    logger.error(error);
-  }
-};
 
 module.exports = {
-  assignAdviser,
   getAllAdvisers,
 };
