+++ back/controllers/admin/student.js
@@ -75,8 +75,40 @@ const getStudent = async (req = request, res = response) => {
     logger.error(error);
   }
 };
+const assignAdviser = async (req, res) => {
+  try {
+    const { idAdviser } = req.body;
+    const { id } = req.params;
+    const student = await StudentModel.update(
+      {
+        adviserId: idAdviser,
+      },
+      {
+        where: {
+          id,
+        },
+      },
+    );
+    if (student < 1) {
+      logger.warn(`ID Student: '${id}', not found in db. Method: PUT.  Url: ${req.originalUrl}.`);
+      return res.status(204).json({
+        message: 'Student not found',
+        data: '',
+      });
+    }
+    res.status(201).json({
+      message: `Adviser assigned to the student with the id ${id}`,
+    });
+  } catch (error) {
+    res.status(500).json({
+      message: 'Server Error',
+    });
+    logger.error(error);
+  }
+};
 module.exports = {
   addStudent,
+  assignAdviser,
   getAllStudent,
   getStudent,
 };
